// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const JenisKelamin = {
  LAKI_LAKI: 'LAKI_LAKI',
  PEREMPUAN: 'PEREMPUAN'
} as const

export type JenisKelamin = (typeof JenisKelamin)[keyof typeof JenisKelamin]


export const AccountRole = {
  ANGGOTA: 'ANGGOTA',
  ADMIN: 'ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN'
} as const

export type AccountRole = (typeof AccountRole)[keyof typeof AccountRole]


export const AccountStatus = {
  AKTIF: 'AKTIF',
  NONAKTIF: 'NONAKTIF',
  MENUNGGU: 'MENUNGGU',
  DITOLAK: 'DITOLAK'
} as const

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus]


export const JenisSimpanan = {
  POKOK: 'POKOK',
  WAJIB: 'WAJIB',
  SUKARELA: 'SUKARELA'
} as const

export type JenisSimpanan = (typeof JenisSimpanan)[keyof typeof JenisSimpanan]


export const StatusSimpanan = {
  MENUNGGU: 'MENUNGGU',
  TERVERIFIKASI: 'TERVERIFIKASI',
  DITOLAK: 'DITOLAK'
} as const

export type StatusSimpanan = (typeof StatusSimpanan)[keyof typeof StatusSimpanan]


export const TipeBungaPinjaman = {
  MENURUN: 'MENURUN',
  FLAT: 'FLAT'
} as const

export type TipeBungaPinjaman = (typeof TipeBungaPinjaman)[keyof typeof TipeBungaPinjaman]


export const StatusPinjaman = {
  MENUNGGU: 'MENUNGGU',
  DISETUJUI: 'DISETUJUI',
  DITOLAK: 'DITOLAK'
} as const

export type StatusPinjaman = (typeof StatusPinjaman)[keyof typeof StatusPinjaman]


export const StatusPembayaranPinjaman = {
  MENUNGGU: 'MENUNGGU',
  TERVERIFIKASI: 'TERVERIFIKASI',
  DITOLAK: 'DITOLAK'
} as const

export type StatusPembayaranPinjaman = (typeof StatusPembayaranPinjaman)[keyof typeof StatusPembayaranPinjaman]